refactor: return Optional<Customer> instead of Customer in getByMail and getByUsername in CustomerRepository.java and CustomerService.java

+++ src/main/java/com/vodafone/service/CartService.java
@@ -29,7 +29,6 @@ public class CartService {
 
 
     public boolean update(Long id, Cart updatedEntity) {
-//        if (updatedEntity == null || get(id) == null)
         if (id == null)
             throw new NullIdException("Null cart id is provided");
         if (updatedEntity == null)
@@ -39,8 +38,8 @@ public class CartService {
 
 
     public boolean delete(Long id) {
-//        if (get(id) == null)
-//          throw new NullPointerException("No cart exists with this id");
+        if (get(id) == null)
+          throw new NullPointerException("No cart exists with this id");
         if (id == null)
             throw new NullIdException("Null cart id is provided");
         return cartRepository.delete(id);
